requirements/base.txt
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-api-python-client==1.7.6

